Replace hardcoded car info with db table

+++ section-2/step-04/multi-agent-system/src/main/java/com/carmanagement/agentic/agents/CarWashAgent.java
@@ -33,7 +33,7 @@ public interface CarWashAgent {
             String carMake,
             String carModel,
             Integer carYear,
-            Integer carNumber,
+            Long carNumber,
             String carWashRequest);
 }
