+++ section-2/step-04/remote-a2a-agent/src/main/java/com/demo/DispositionTool.java
@@ -43,7 +43,7 @@ public class DispositionTool {
      */
     @Tool(name = "DispositionTool")
     public String requestDisposition(
-            Integer carNumber,
+            Long carNumber,
             String carMake,
             String carModel,
             Integer carYear,
